+++ common/src/main/java/com/oingmaryho/business/common/entity/BaseEntity.java
@@ -1,6 +1,7 @@
 package com.oingmaryho.business.common.entity;
 
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -41,4 +42,27 @@ public abstract class BaseEntity {
 
 	@LastModifiedBy
 	private Long updatedBy;
+
+	@Column(nullable = false)
+	@Builder.Default
+	private Boolean isDeleted = false;
+
+	/**
+	 * 엔티티 soft delete
+	 * @param deleteUserId 삭제자 id
+	 */
+	public void softDelete(Long deleteUserId) {
+		this.deletedAt = LocalDateTime.now();
+		this.deletedBy = deleteUserId;
+		this.isDeleted = true;
+	}
+
+	/**
+	 * 엔티티 soft delete 취소
+	 */
+	public void restore() {
+		this.deletedAt = null;
+		this.deletedBy = null;
+		this.isDeleted = false;
+	}
 }
