+++ delivery-service/src/main/java/com/oingmaryho/business/delivery_service/domain/DeliveryManager.java
@@ -1,5 +1,6 @@
 package com.oingmaryho.business.delivery_service.domain;
 
+import com.oingmaryho.business.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
@@ -16,7 +17,7 @@ import java.util.UUID;
 @DynamicUpdate
 @Table(name = "p_delivery_manager")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-class DeliveryManager extends BaseEntity{
+class DeliveryManager extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
